pyproject.toml
[project]
name = "blender-extension-template"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "lokimckay", email = "lokimckay@gmail.com"},
]
dependencies = []
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}

[dependency-groups]
dev = [
    "platformdirs>=4.3.6",
    "pyyaml>=6.0.2",
    "fake-bpy-module>=20241230",
]

[tool.pdm]
distribution = true

[tool.pdm.scripts]
dev = "python scripts/dev.py"
clean = "rm -rf build"
build = "python scripts/build.py"
